+++ src/cli.rs
@@ -73,6 +73,10 @@ pub fn run() -> anyhow::Result<()> {
         }
     }
 
+    if let Some(config) = Config::from_env("BACON_PREFS")? {
+        settings.apply_config(&config);
+    }
+
     let location = MissionLocation::new(&args)?;
     info!("mission location: {:#?}", &location);
 
@@ -99,6 +103,10 @@ pub fn run() -> anyhow::Result<()> {
     }
     if package_config_path.exists() {
         let config = Config::from_path(&package_config_path)?;
+        settings.apply_config(&config);
+    }
+
+    if let Some(config) = Config::from_env("BACON_CONFIG")? {
         settings.apply_config(&config);
     }
 
@@ -129,7 +137,7 @@ pub fn run() -> anyhow::Result<()> {
     let mut result = Ok(());
     #[allow(clippy::while_let_loop)]
     loop {
-        let (job_name, job) = match job_stack.pick_job(&next_job) {
+        let (concrete_job_ref, job) = match job_stack.pick_job(&next_job) {
             Err(e) => {
                 result = Err(e);
                 break;
@@ -139,7 +147,7 @@ pub fn run() -> anyhow::Result<()> {
                 break;
             }
         };
-        let r = Mission::new(&location, job_name.to_string(), job, &settings)
+        let r = Mission::new(&location, concrete_job_ref, job, &settings)
             .and_then(|mission| app::run(&mut w, mission, &event_source));
         match r {
             Ok(Some(job_ref)) => {