+++ src/job_stack.rs
@@ -6,8 +6,8 @@ use {
     },
 };
 
-/// The stack of jobs that bacon ran, allowing
-/// to get back to the previous one
+/// The stack of jobs that bacon ran, allowing to get back to the previous one,
+/// or to scope the current one
 pub struct JobStack<'c> {
     settings: &'c Settings,
     entries: Vec<ConcreteJobRef>,
@@ -28,6 +28,9 @@ impl<'c> JobStack<'c> {
             .unwrap_or(&self.settings.default_job)
     }
 
+    /// Apply the job ref instruction to determine the job to run, updating the stack.
+    ///
+    /// When no job is returned, the application is supposed to quit.
     pub fn pick_job(
         &mut self,
         job_ref: &JobRef,
@@ -37,19 +40,38 @@ impl<'c> JobStack<'c> {
             JobRef::Default => self.settings.default_job.clone(),
             JobRef::Initial => self.initial_job().clone(),
             JobRef::Previous => {
-                self.entries.pop();
+                let current = self.entries.pop();
                 match self.entries.pop() {
                     Some(concrete) => concrete,
+                    None if current
+                        .as_ref()
+                        .map_or(false, |current| current.scope.has_tests()) =>
+                    {
+                        // rather than quitting, we assume the user wants to "unscope"
+                        ConcreteJobRef {
+                            name_or_alias: current.unwrap().name_or_alias,
+                            scope: Scope::default(),
+                        }
+                    }
                     None => {
                         return Ok(None);
                     }
                 }
             }
             JobRef::Concrete(concrete) => concrete.clone(),
+            JobRef::Scope(scope) => match self.entries.last() {
+                Some(concrete) => ConcreteJobRef {
+                    name_or_alias: concrete.name_or_alias.clone(),
+                    scope: scope.clone(),
+                },
+                None => {
+                    return Ok(None);
+                }
+            },
         };
-        let job = match &concrete {
-            ConcreteJobRef::Alias(alias) => Job::from_alias(alias, self.settings),
-            ConcreteJobRef::Name(name) => self
+        let job = match &concrete.name_or_alias {
+            NameOrAlias::Alias(alias) => Job::from_alias(alias, self.settings),
+            NameOrAlias::Name(name) => self
                 .settings
                 .jobs
                 .get(name)